% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{default_repo}
\alias{default_repo}
\title{Default SVN Repository}
\usage{
default_repo(serverip = getOption("svn_serverip"),
  root_dir = "svn/kwb")
}
\arguments{
\item{serverip}{serverip for repo (default: getOption("svn_serverip") )}

\item{root_dir}{root directory (default: "svn/kwb")}
}
\value{
string to svn repo with login
}
\description{
Default SVN Repository
}
