% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{extractCommitInfo}
\alias{extractCommitInfo}
\title{Get Commit Infos}
\usage{
extractCommitInfo(commits)
}
\arguments{
\item{commits}{commits}
}
\value{
commit infos
}
\description{
Get Commit Infos
}
