% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{extractEntryNameAndCommit}
\alias{extractEntryNameAndCommit}
\title{Extract Entry Name and Commit}
\usage{
extractEntryNameAndCommit(entryName, entryCommit)
}
\arguments{
\item{entryName}{entryName}

\item{entryCommit}{entryCommit}
}
\value{
data.frame with name, author and data
}
\description{
Extract Entry Name and Commit
}
