% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{get_rscript_paths}
\alias{get_rscript_paths}
\title{Default SVN Repository RScripts Paths}
\usage{
get_rscript_paths()
}
\arguments{
\item{root_dir}{root directory (default: "svn/kwb/R_Development/trunk/R")}
}
\description{
Default SVN Repository RScripts Paths
}
