% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{getRepoInfo}
\alias{getRepoInfo}
\title{Get Repo Info}
\usage{
getRepoInfo(currentRevision = 1920, startRevision = 2,
  tDir = tempdir(), repo = default_repo(), logs = TRUE, dbg = TRUE)
}
\arguments{
\item{currentRevision}{latest SVN revision}

\item{startRevision}{start SVN revision}

\item{tDir}{target directory (default: tempdir())}

\item{repo}{repository (default: default_repo())}

\item{logs}{if TRUE logs info is extracted, if FALSE size info (default: TRUE)}

\item{dbg}{debug messages (default: TRUE)}
}
\value{
writes files with repo log info into target directory and returns 
path to target directory
}
\description{
Get Repo Info
}
