% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{getRevisionInfo}
\alias{getRevisionInfo}
\title{Get Revision Information}
\usage{
getRevisionInfo(
  revision,
  tDir = tempdir(),
  repo = default_repo(),
  logs = TRUE,
  fPrefix = "ente",
  dbg = TRUE
)
}
\arguments{
\item{revision}{revision number}

\item{tDir}{target directory}

\item{repo}{repository}

\item{logs}{if TRUE logs info is extracted, if FALSE size info (default:
TRUE)}

\item{fPrefix}{file prefix (default: "ente")}

\item{dbg}{debug messages (default: TRUE)}
}
\value{
revision infos
}
\description{
Get Revision Information
}
