% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_histories.R
\name{init_revision_summary}
\alias{init_revision_summary}
\title{Initialise Revision Summary}
\usage{
init_revision_summary(hist_log, file, path, start, end)
}
\description{
Initialise Revision Summary
}
\keyword{internal}
