% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_histories.R
\name{read_histories}
\alias{read_histories}
\title{Read Histories}
\usage{
read_histories(history_dir)
}
\arguments{
\item{history_dir}{directory with history files (as created with}
}
\value{
data.frame with columns
}
\description{
Read Histories
}
