% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{readLogFilesInDirectory}
\alias{readLogFilesInDirectory}
\title{Read SVN Logfiles In A Directory}
\usage{
readLogFilesInDirectory(logdir, pattern = "_log_")
}
\arguments{
\item{logdir}{directory containing svn log information}

\item{pattern}{pattern (default: "_log_")}
}
\value{
data frame with svn log data
}
\description{
Read SVN Logfiles In A Directory
}
