% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{readSizeFiles}
\alias{readSizeFiles}
\title{Read Size Files}
\usage{
readSizeFiles(fDir, fPattern = "ente_r", dbg = FALSE)
}
\arguments{
\item{fDir}{directory containing data with file size information}

\item{fPattern}{file pattern (default: "ente_r")}

\item{dbg}{debug (default: FALSE)}
}
\value{
list with size file information
}
\description{
Read Size Files
}
