% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{default_repo}
\alias{default_repo}
\title{Default SVN Repository}
\usage{
default_repo(root_dir = "svn/kwb",
  username = getOption("svn_username"),
  password = getOption("svn_password"),
  serverip = getOption("svn_serverip"))
}
\arguments{
\item{root_dir}{root directory (default: "svn/kwb")}

\item{username}{username for repo (default: getOption("svn_username") )}

\item{password}{password for repo (default: getOption("svn_password") )}

\item{serverip}{serverip for repo (default: getOption("svn_serverip") )}
}
\value{
string to svn repo with login
}
\description{
Default SVN Repository
}
