% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{default_repo}
\alias{default_repo}
\title{Default SVN Repository}
\usage{
default_repo(
  serverip = Sys.getenv("SVN_IP"),
  username = Sys.getenv("SVN_USER"),
  password = Sys.getenv("SVN_PW"),
  root_dir = "svn/kwb"
)
}
\arguments{
\item{serverip}{serverip for repo (default: Sys.getenv("SVN_IP") )}

\item{username}{username for repo (default: Sys.getenv("SVN_USER"))}

\item{password}{password for repo (default: Sys.getenv("SVN_PW"))}

\item{root_dir}{root directory (default: "svn/kwb")}
}
\value{
string to svn repo with login
}
\description{
Default SVN Repository
}
