% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logfiles.R
\name{get_rscript_paths}
\alias{get_rscript_paths}
\title{Default SVN Repository RScripts Paths}
\usage{
get_rscript_paths(rscripts_path = default_rscripts())
}
\arguments{
\item{rscripts_path}{default R scripts path default: \code{\link{default_rscripts}}}
}
\description{
Default SVN Repository RScripts Paths
}
