% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_history.R
\name{read_files_history}
\alias{read_files_history}
\title{Read SVN History For Files}
\usage{
read_files_history(repo = default_rscripts(), tdir = tempdir(), dbg = TRUE)
}
\arguments{
\item{repo}{path to repository (default: default_repo())}

\item{tdir}{target directory (default: tempdir())}

\item{dbg}{debug (default: TRUE)}
}
\value{
writes history files to target directory and returns target directory
}
\description{
Read SVN History For Files
}
